P1: make server authoritative for audit identity
